// src/app/shared/models/enums/route-names.ts
export enum RouteNames {
  Dashboard = 'Dashboard',
  SendMessage = 'New Communication',
  Groups = 'Groups/Members',
  MessageResults = 'Communication Results',
  MessageLibrary = 'Message Library',
  CallInSettings = 'Call-In Settings',
  RecurringMessage = 'My Recurring Message',
  Keywords = 'Text Keywords',
  assignKeyword = 'Assign Keyword to Group',
  CustomPhone = 'Custom Phone Numbers',
  Reminders = 'Auto Reminders',
  Rewards = 'Rewards Center',
  RewardHistory = 'Reward History',
  GroupManagers = 'Group Managers',
  System = 'System Settings',
  HelpCenter = 'Help Center',
  InformationFifty = 'Get $50',
  User = 'Profile',
  BillingPlan = 'Billing & Plan Details',
  Logout = 'Logout',
  MemberProfile = 'Member Profile',
  PauseAccount = 'Pause Account',
  BillingHistory = 'History',
  BillingInvoice = 'Invoice',
  PasswordLogin = 'Password/Login',
  PlanDetails = 'Plan Details',
  EditMessage = 'Edit Message',
  SwitchAccount = 'Switch Account',
  ChangePlan = 'Change Plan',
  AddCredits = 'Add Credits',
  Notifications = 'Notifications',
}